update and fix bug

+++ firmware/main/ucfg.h
@@ -47,7 +47,7 @@ bool UCFG_write_temp_limit(uint8_t value);
 bool UCFG_read_temp_limit(uint8_t* value);
 bool UCFG_write_connection(uint8_t value);
 bool UCFG_read_connection(uint8_t* value);
-bool UCFG_write_mqtt_ca(uint8_t* data, uint8_t len);
+bool UCFG_write_mqtt_ca(uint8_t* data, uint16_t len);
 bool UCFG_read_mqtt_ca(uint8_t* data, uint16_t* len);
 
 void UCFG_test(void);
